Solve comment about tests
