+++ spec/dummy/app/helpers/test_helper.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+class TestHelper < ActionView::Base; end
