+++ spec/form_builder/currency_fields_spec.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Bali::FormBuilder do
+  include_context 'form builder'
+
+  describe '#currency_field_group' do
+    it 'renders an input' do
+      expect(builder.currency_field_group(:budget)).to include(
+        '<div id="field-budget" class="field-group-wrapper-component field ">'\
+        '<label class="label " for="movie_budget">Budget</label>'\
+        '<div class="field has-addons"><div class="control">'\
+        '<span class="button is-static">$</span></div>'\
+        '<div class="control "><input placeholder="0" step="0.01" '\
+        'pattern="^(\\d+|\\d{1,3}(,\\d{3})*)(\\.\\d+)?$" class="input " type="text" '\
+        'name="movie[budget]" id="movie_budget" /></div></div></div>'
+      )
+    end
+  end
+end
