+++ spec/support/shared_context/form_builder.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.shared_context 'form builder', shared_context: :metadata do
+  let(:helper) { TestHelper.new(ActionView::LookupContext.new(ActionView::PathSet.new), {}, nil) }
+  let(:resource) { Movie.new }
+  let(:builder) { Bali::FormBuilder.new :movie, resource, helper, {} }
+end
